16_Bixby/bixby-winary-master/backend/winary/requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.5
django-cors-headers==3.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
idna==2.8
oauthlib==3.1.0
PyJWT==1.7.1
pymongo==3.9.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
sqlparse==0.3.0
urllib3==1.25.6
